+++ rosidl_generator_c/include/rosidl_generator_c/primitives_sequence_functions.h
