+++ rosidl_generator_c/resource/msg__struct.h.em
@@ -51,7 +51,7 @@ for field in spec.fields:
             field_names.append(field.name)
         else:
             if field.type.is_dynamic_array():
-                field_names = includes.setdefault('rosidl_generator_c/primitives_array.h', [])
+                field_names = includes.setdefault('rosidl_generator_c/primitives_sequence.h', [])
                 field_names.append(field.name)
     else:
         field_names = includes.setdefault(
