+++ rosidl_generator_c/src/primitives_sequence_functions.c
