+++ rosidl_generator_c/test/separate_compilation.c
@@ -22,7 +22,7 @@
 #include "rosidl_generator_c/msg/nested__functions.h"
 #include "rosidl_generator_c/msg/various__struct.h"
 #include "rosidl_generator_c/msg/various__functions.h"
-#include "rosidl_generator_c/primitives_array_functions.h"
+#include "rosidl_generator_c/primitives_sequence_functions.h"
 #include "rosidl_generator_c/string_functions.h"
 
 int func()
